# Conflicts:
#	lucene/core/src/test/org/apache/lucene/search/TestTermQuery.java

+++ lucene/highlighter/src/java/org/apache/lucene/search/uhighlight/PassageScorer.java
@@ -34,8 +34,10 @@ public class PassageScorer {
 
   /** BM25 k1 parameter, controls term frequency normalization */
   final float k1;
+
   /** BM25 b parameter, controls length normalization. */
   final float b;
+
   /** A pivot used for length normalization. */
   final float pivot;
 
@@ -110,7 +112,7 @@ public class PassageScorer {
   }
 
   public float score(Passage passage, int contentLength) {
-    float score = 0;
+    double score = 0d;
     BytesRefHash termsHash = new BytesRefHash();
     int hitCount = passage.getNumMatches();
     int[] termFreqsInPassage = new int[hitCount]; // maximum size
@@ -132,6 +134,6 @@ public class PassageScorer {
           tf(termFreqsInPassage[i], passage.getLength()) * weight(contentLength, termFreqsInDoc[i]);
     }
     score *= norm(passage.getStartOffset());
-    return score;
+    return (float) score;
   }
 }
